chore(tests): port and re-enable `TestAdmissionWebhook_SecretCredentials`
